more data uploaded
concerning is the figure from Gillenwater et al 1998 that show spectra "representative" of cancer. the paper does not say where the tissue was measured. Porphyrins are commonly found on tongue, but not in buccal mucosa. So, location, location, location matters. Concerning to me is that so many clinical researchers jumped on the bandwagon believing that porphyrins were diagnostic of cancer.  Later, others reported measuring it on healthy tongues