we can remove these notes later
but they were helpful for me to remember one possible thing to try
namely, testing the hypothesis that the only difference between the lip and tongue is keratin (between 500 and 600 nm) and porphyrins and chlorophyll (between 600 and 700 nm).  We do not need to search for the spectral signature of porphyrins and chlorophyll because there is agreement in the literature.  Collagen and keratin emissions change depending on "structural polymerization" in human tissue. the intrinsic peak wavelength of FAD (~520–530 nm) remains relatively stable, so differences in the literature may be due to differences in measurement geometry (lighting, detector, etc).